test(express): split test file
see #120
